automation/tps_monthly_reports/requirements.txt
jaydebeapi
pyyaml
pypdf2
pandas